infra/lambda-functions/monthly_summary/requirements.txt
langchain-core
langchain_community
langchain-openai